+++ cyclonedx-bom/src/specs/common/mod.rs
@@ -25,6 +25,7 @@ pub(crate) mod dependency;
 pub(crate) mod external_reference;
 pub(crate) mod hash;
 pub(crate) mod license;
+pub(crate) mod metadata;
 pub(crate) mod organization;
 pub(crate) mod property;
 pub(crate) mod service;
