+++ cyclonedx-bom/src/specs/v1_3/mod.rs
@@ -18,9 +18,9 @@
 
 pub(crate) mod bom;
 pub(crate) mod composition;
-pub(crate) mod metadata;
 
 pub(crate) use crate::specs::common::component::v1_3 as component;
 pub(crate) use crate::specs::common::external_reference::v1_3 as external_reference;
+pub(crate) use crate::specs::common::metadata::v1_3 as metadata;
 pub(crate) use crate::specs::common::service::v1_3 as service;
 pub(crate) use crate::specs::common::tool::v1_3 as tool;
